Move constants to files where they are used; Config to buffer
