+++ tracing/ops_tracing/_api.py
@@ -21,7 +21,7 @@ import logging
 
 import ops
 
-from ._const import Config
+from ._buffer import Config
 from .vendor.charms.certificate_transfer_interface.v1.certificate_transfer import (
     CertificateTransferRequires,
 )
