+++ tracing/test/test_storage.py
@@ -21,7 +21,7 @@ import ops.testing
 import pytest
 
 from ops_tracing import _backend
-from ops_tracing._const import Config
+from ops_tracing._buffer import Config
 
 _pydantic = pytest.importorskip('pydantic')
 _export = pytest.importorskip('ops._tracing.export')
